+++ v3/src/components/case-table/use-index-column.tsx
@@ -1,9 +1,10 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react"
 import { createPortal } from "react-dom"
-import { Menu, MenuButton, MenuItem, MenuList } from "@chakra-ui/react"
+import { Menu, MenuButton } from "@chakra-ui/react"
 import { IDataSet } from "../../data-model/data-set"
 import { kIndexColumnKey, TColumn, TFormatterProps } from "./case-table-types"
 import { ColumnHeader } from "./column-header"
+import { IndexMenuList } from "./index-menu-list"
 
 interface IHookProps {
   data?: IDataSet
@@ -61,46 +62,15 @@ export const IndexCell = ({ caseId, index, onClick }: ICellProps) => {
 
   // Find the parent CODAP component to display the index menu above the grid
   useEffect(() => {
-    if (cellElt && !codapComponentElt) {
-      let parent: HTMLElement | null
-      for (parent = cellElt; parent; parent = parent.parentElement) {
-        if (parent.classList.contains("codap-component")) {
-          setCodapComponentElt(parent)
-          break
-        }
-      }
-    }
-  }, [cellElt, codapComponentElt])
-
-  const handleMoveDataRow = () => {
-    alert("Move Data Row Here clicked")
-  }
-
-  const handleInsertCase = () => {
-    alert("Insert Case clicked")
-  }
-
-  const handleInsertCases = () => {
-    alert("Insert Cases clicked")
-  }
-
-  const handleDeleteCase = () => {
-    alert("Delete Case clicked")
-  }
+    setCodapComponentElt(cellElt?.closest(".codap-component") as HTMLDivElement ?? null)
+  }, [cellElt])
 
   return (
     <Menu isLazy>
       <MenuButton ref={setNodeRef} className="codap-index-content" data-testid="codap-index-content">
         {index != null ? `${index + 1}` : ""}
       </MenuButton>
-      {codapComponentElt && createPortal((
-        <MenuList>
-          <MenuItem onClick={handleMoveDataRow}>Move Data Entry Row Here</MenuItem>
-          <MenuItem onClick={handleInsertCase}>Insert Case</MenuItem>
-          <MenuItem onClick={handleInsertCases}>Insert Cases...</MenuItem>
-          <MenuItem onClick={handleDeleteCase}>Delete Case</MenuItem>
-        </MenuList>
-      ), codapComponentElt)}
+      {codapComponentElt && createPortal(<IndexMenuList caseId={caseId} index={index}/>, codapComponentElt)}
     </Menu>
   )
 }
